requirements.txt
langchain==0.0.322
chromadb==0.3.29
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
numpy==1.26.4
pandas==2.3.3
hnswlib==0.8.0
python-dotenv==1.2.1
requests==2.32.5
uvicorn==0.18.3
fastapi==0.85.1
huggingface-hub==0.25.2

